API client refactoring
