+++ lib/contentful/apiClient.ts
@@ -1,13 +1,10 @@
 import { createClient } from "contentful";
 
-const space = process.env.NEXT_PUBLIC_SPACE;
-const accessToken = process.env.NEXT_PUBLIC_ACCESS_TOKEN;
-
-if (!space || !accessToken) {
-  throw new Error("Space or Access Token is missing in environment variables.");
-}
-
-export const client = createClient({
-  space,
-  accessToken,
+export const contentfulClient = createClient({
+  accessToken: process.env.CONTENTFUL_ACCESS_TOKEN!,
+  space: process.env.CONTENTFUL_SPACE!,
 });
+
+export async function getEntityData<T>(entryId: string) {
+  return contentfulClient.getEntry<T>(entryId);
+}
